Add `entrypoints.py` as a place to construct the app
Prior to this change, we had no common base level location in the
project to construct the API and orchestrate the various levels of the
application together.

This change adds the `entrypoints.py` file which can be used to glue
together things from various levels of the application and expose to
built application.