Add basic domain config
Prior to this change, there was no way to get a public and private key
pair.

This change adds some simple methods for creating a niave set of keys on
each launch, it will need to be updated later to load the `_KEY_PATH`
from an env variable but for now this should be fine to generate a new
key on each launch. This commit also includes the relevant tests.