Update `config.py` to use `tempfile.mkdtemp()`
Prior to this change, each test run was using the same temporary folder.

This change updates that so that the config will return a new tempdir
each run. This will need to be updated one day to use a secure secret
storage service for these certs rather than generating them on each run.