Add dev dependancy `pytest-testmon`
Prior to this change, we had no way to run only the tests that are
required to be re ran due to edits on the source code.

This change adds `pytest-testmon` which uses coverage to work out which
tests cover which lines in each source code file. This allows these
tests to be reran only for the files that have been edited.